tried a fix for too long run: int -> long for iteration counter

+++ include/triangulation.h
@@ -104,8 +104,8 @@ public: /// @todo deve essere private
 public:
     // DI SERVIZIO (misurano gli step della simulazione)
     
-    int volume_step;
-    int steps_done;
+    long volume_step;
+    long steps_done;
     long iterations_done;
     
     // ##### STARTING TRIANGULATION INITIALIZATION #####
